requirements.txt
Django==2.0.2
django-debug-toolbar==1.9.1
djangorestframework==3.7.7
Pillow==5.0.0
pytz==2017.3
selenium==3.11.0
sqlparse==0.2.4
